46 commits staged and bench-validated together rather than landed piecemeal,
so cross-cluster interaction was caught on one build instead of on the release
line. integration was 46 ahead / 0 behind, so this merge is conflict-free.

What ships:

- Stitching audit Phases 1/2/4 -- one canonical TIFF predicate (OME tiles were
  double-counted), pixel size read from the tile's PhysicalSizeX instead of
  re-derived at binning=1, center metadata from the extent midpoint, shared-range
  normalize across the seam, producing algorithm stamped into output metadata,
  and the dead _position_stitcher path deleted with its tests repointed.
- Depth-contract Phase 0 -- the 8-bit downconvert is now explicit and typed.
- CPU/GPU perf -- MSAA off, the redundant per-frame Pylon copy dropped,
  GPU-minify for the live preview (host downscale opt-in), and the
  out-of-process py-spy profiler with its ground-truth check.
- TIFF metadata -- tile metadata recovered from the ImageJ container; the
  stitcher now warns by name when tiles carry no pixel size.

Issue fixes, each verified at the bench on this build:

- #746 12-bit false-color composites render as 8-bit RGB, not near-black.
- #717 the live crosshair stays visible through a protocol run.
- #424 the manual-video manifest ignores None-padded and skipped frame slots.
- #745 the redundant Stitch panel paragraph is gone; tooltips already covered it.

One revert rides along: a0f80ec1 (settings validation) is backed out by 8ecf113d.
It had converted two survivable states -- settings.json absent, and unparseable --
into a failure to launch, against a loader whose established design is degrade
rather than die. Its interlock test could not detect the condition it existed to
prevent, because the file opened at runtime is the per-user copy, not the
repository's. The original item is dispositioned decide-not-to-fix: the bare
subscript it targeted is only reachable on a broken install, and the per-user
data directory is version-scoped, so each beta bump re-copies the key.

Depth and perf items B3-B6 are validated by field use over several weeks rather
than re-run on this trip.

Bench: all six trip-sheet checks green (B1, B2, #746, #717, #424, #745).
Tests: full LVP suite green at a21664d8; ruff check and format --check both clean.

+++ modules/stitch_algorithms.py
@@ -335,8 +335,8 @@ def estimate_phase_offset(
         mov_view.astype(np.float32),
         window,
     )
-    corr_x = int(round(-shift[0]))
-    corr_y = int(round(-shift[1]))
+    corr_x = round(-shift[0])
+    corr_y = round(-shift[1])
     if abs(corr_x) > max_correction_px or abs(corr_y) > max_correction_px:
         return 0, 0, float(response)
     return corr_x, corr_y, float(response)
